pyproject.toml
[project]
name = "learn-rabbitmq"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery[sqs]>=5.6.2",
    "fastapi[standard]>=0.129.0",
    "kombu>=5.6.2",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "redis>=7.2.0",
]

[dependency-groups]
dev = [
    "pyclean>=3.5.0",
]
